dhcrelay: make a note about the intf->index == 0 bug
This baffles me a bit that the system would allow intf entries
to remain without a valid index because the Layer 2 parsing
failed.  In reality that doesn't happen when the types are
properly set but that also probably was the reason for:

https://github.com/openbsd/src/commit/9a58347058de

So when introducing a new type this fails pretty badly.  Since
we don't know what FreeBSD uses it may be better to remove the
entries before returning.