dhcrelay: remove faulty OpenBSD code that broke index reads
Noticed on FreeBSD dhclient dispatch.c that index was always
correctly set and it turns out in 2017 this broke by adding
the IFT exceptions discussed over at OpenBSD lately but ultimately
discarded the notion of it being suboptimal:

https://github.com/openbsd/src/commit/962cae8e7a

This restores the index parsing as it was originally done but
then has to postprocess anyway in order to sanity check the
hardware address required.  Here we loosely check ETHER_ADDR_LEN
for now, better than taking zero-length addresses.

This also brings up bridges but removes enc support which
is probably more academic than used out in the wild (and
only for IPv4?).

https://github.com/openbsd/src/commit/4be048dca

We can improve the sanity checking later. For now I'm happy
to transition out of a "the other side does it wrong" situation
sabotaging the interface list.

Another question is why the daemon needs the full interface list
and in the case of dhcrelay6 triggers multicast join on all the
interfaces instead of just the one(s) selected on the command line.
Probably a can of works for later.